feat: se añade crud para insumo y otros metodos de busqueda, venta y calculo.

+++ src/main/java/com/burgerreal/burger_gestion/infrastructure/adapter/in/web/dto/venta/VentaDetalleResponse.java
@@ -4,16 +4,17 @@ import com.burgerreal.burger_gestion.domain.enums.EstadoVenta;
 import com.burgerreal.burger_gestion.infrastructure.adapter.in.web.dto.metodo_pago.MetodoPagoDetalleParaVentaResponse;
 import com.fasterxml.jackson.annotation.JsonFormat;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record VentaDetalleResponse(
         Long id,
         @JsonFormat(pattern = "dd-MM-yyyy HH:mm:ss")
         LocalDateTime fecha,
-        Long montoTotalBruto,
-        Long costoTotalInsumos,
-        Long comisionPasarela,
-        Long gananciaNeta,
+        BigDecimal montoTotalBruto,
+        BigDecimal costoTotalInsumos,
+        BigDecimal comisionPasarela,
+        BigDecimal gananciaNeta,
         boolean pagoConfirmado,
         Long cargoPorAnulacionCocina,
         EstadoVenta estadoVenta,
